+++ TrackPropagation/RungeKutta/src/CartesianStateAdaptor.h
@@ -4,7 +4,7 @@
 class dso_internal CartesianStateAdaptor {
 public:
 
-    typedef double                            Scalar;
+    typedef float                             Scalar;
     typedef Basic3DVector<Scalar>             Vector3D;
 
     CartesianStateAdaptor( const RKSmallVector<double,6>& rk) :
